// src/app/kyc/models/userKYCUrlOutputModel.ts
/**
 * OMRO DigitalApp API
 * API used on OMRO Digital Resources
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { KYCUrlType } from './kYCUrlType';

/**
 * URL and metadata to show the KYC screens
 */
export interface UserKYCUrlOutputModel { 
    /**
     * Type of KYC step
     */
    type?: KYCUrlType;
    /**
     * URL to be used in IFRAME
     */
    url?: string;
    /**
     * Title for the modal preceding the IFRAME
     */
    title?: string;
    /**
     * Details for the modal preceding the IFRAME
     */
    details?: string;
    /**
     * Button text for the modal preceding the IFRAME
     */
    buttonText?: string;
}